requirements.txt
flask
flask-cors
psycopg2-binary
pymongo
openai
python-dotenv
